refactor: Remove exported bot const
