+++ src/telegram/bot/index.ts
@@ -1,14 +1,24 @@
 import { Dispatcher, filters } from "@mtcute/dispatcher"
-import { bot } from "#telegram/bot/bot"
 import { downloadDp } from "#telegram/bot/download"
 import { settingsDp } from "#telegram/bot/settings"
 import { statsDp } from "#telegram/bot/stats"
 import { translatorFor } from "#telegram/helpers/i18n"
-import { handleDispatcherError } from "#telegram/bot/errors"
+import { createDispatcherErrorHandler } from "#telegram/bot/errors"
+import { TelegramClient } from "@mtcute/node"
+import { env } from "#telegram/helpers/env"
 
-export const dp = Dispatcher.for(bot)
+const bot = new TelegramClient({
+    apiId: env.API_ID,
+    apiHash: env.API_HASH,
+    storage: "data/session",
+    updates: {
+        catchUp: true,
+    },
+})
 
-dp.onError(handleDispatcherError)
+const dp = Dispatcher.for(bot)
+
+dp.onError(createDispatcherErrorHandler(bot))
 dp.onNewMessage(filters.command("start"), async (msg) => {
     const t = await translatorFor(msg.sender)
     await msg.replyText(t("start"))
@@ -18,4 +28,9 @@ dp.extend(settingsDp)
 dp.extend(statsDp)
 dp.extend(downloadDp)
 
-export { startBot } from "#telegram/bot/bot"
+export const startBot = async () => {
+    bot.run({ botToken: env.BOT_TOKEN }, async () => {
+        if (env.ERROR_CHAT_ID)
+            await bot.sendText(env.ERROR_CHAT_ID, "started bot :з")
+    })
+}
